+++ src/client/gui/src/main/java/com/kdyncs/dragonsong/client/gui/components/MainView.java
@@ -1,7 +1,18 @@
 package com.kdyncs.dragonsong.client.gui.components;
 
+import net.rgielen.fxweaver.core.FxWeaver;
 import net.rgielen.fxweaver.core.FxmlView;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
 
-@FxmlView
+@Component
+@FxmlView("/view/MainView.fxml")
 public class MainView {
+
+    private final FxWeaver weaver;
+
+    @Autowired
+    public MainView(FxWeaver weaver) {
+        this.weaver = weaver;
+    }
 }
