+++ src/client/gui/src/main/java/com/kdyncs/dragonsong/client/gui/event/StageReadyEvent.java
@@ -1,0 +1,14 @@
+package com.kdyncs.dragonsong.client.gui.event;
+
+import javafx.stage.Stage;
+import org.springframework.context.ApplicationEvent;
+
+public class StageReadyEvent extends ApplicationEvent {
+
+    public final Stage stage;
+
+    public StageReadyEvent(Stage stage) {
+        super(stage);
+        this.stage = stage;
+    }
+}
