+++ src/client/gui/src/main/resources/view/MainView.fxml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.text.Text?>
+
+
+<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/17" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.kdyncs.dragonsong.client.gui.components.MainView">
+   <children>
+      <Text layoutX="96.0" layoutY="109.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Hello World" wrappingWidth="129.13671875" />
+   </children>
+</AnchorPane>
